LB Routes should be distributed back to the originator
As a LoadBalancer can exist together with a NetworkInterface
on the same Hypervisor and share the same metalbond client,
the LB announcement of the NetworkInterface should be reflected
back to originator too. Otherwise a possible LoadBalancer
on the same Hypervisor node doesnt have a chance to receive
the LB announcement of that NetworkInterface.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>